Printscreen helper script for slurp/grim
Forks off slurp (program to select a region on the screen), then inserts a click event (to start region selection), once slurp finishes gets the output from a named pipe and cats it to std out.
Planning on using this to have print screening be more streamlined. Was also a good learning exercise on named pipes, forking of processes, silencing various std outputs and the differences between different shells when it comes to forking and output.